submit-to-bigquery-function/requirements.txt
functions-framework==3.*
Flask-Cors==5.0.0
Flask==3.0.3
google-cloud==0.34.0
protobuf==5.28.2
google-cloud-bigquery==3.26.0
python-dateutil==2.9.0.*